Add primary email verification status
