+++ lib/zodSchemas.ts
@@ -1,11 +1,11 @@
 import { z } from "zod";
 
-export const userInformationSchema = z.object({
+export const profileInformationSchema = z.object({
   firstName: z.string().nonempty({ message: "First name cannot be blank" }),
   lastName: z.string().nonempty({ message: "Last name cannot be blank" }),
   username: z.string().nonempty({ message: "Username cannot be blank" }),
 });
 
-export const userEmailSchema = z.object({
+export const emailAddressSchema = z.object({
   emailAddress: z.string().email({ message: "Must be a valid email " }),
 });
